Moving to single handler per event + test scaffolding for search and cache in handlers

+++ shiftcontent/event_handlers/content_item_create.py
@@ -1,6 +1,8 @@
 from shiftevent.handlers.base import BaseHandler
 from shiftcontent.item import Item
 from shiftcontent import db
+from shiftcontent import search_service
+from shiftcontent import cache_service
 from pprint import pprint as pp
 
 
@@ -34,12 +36,20 @@ class ContentItemCreate(BaseHandler):
         :param event: shiftcontent.events.event.Event
         :return: shiftcontent.events.event.Event
         """
+        # create
         item = Item(**event.payload)
-
         items = db.tables['items']
         with db.engine.begin() as conn:
             result = conn.execute(items.insert(), **item.to_db(update=False))
-            item.id = result.inserted_primary_key[0]
+            item.set_field(
+                'id', initial=True, value=result.inserted_primary_key[0]
+            )
+
+        # cache
+        cache_service.set(item)
+
+        # index
+        search_service.put_to_index(item)
 
         return event
 
@@ -50,11 +60,23 @@ class ContentItemCreate(BaseHandler):
         :param event: shiftcontent.events.event.Event
         :return: shiftcontent.events.event.Event
         """
+        from shiftcontent import content_service
+        item = content_service.get_item(event.object_id)
+        if not item:
+            return event
+
+        # delete
         items = db.tables['items']
         with db.engine.begin() as conn:
             query = items.delete()\
                 .where(items.c.object_id == event.object_id)
             conn.execute(query)
+
+        # remove from cache
+        cache_service.delete(item.object_id)
+
+        # remove from index
+        search_service.delete(item.type, item.object_id)
 
         return event
